Add cert-based API server authentication
Add the ability to authenticate incoming requests, verifying that all
requests originate from the Kubernetes API server and no where else.

Authenticating the API server requires manual steps to configure both
the API server and the webhook. Follow the Kubernetes webhook
documentation[1] to create an admission configuration and kubeconfig for
the API server. Only cert-based authentication is supported, basic auth
and token authentication will not be recognized. Then, set clientCA in
the webhook chart's values.yaml to the base64-encoded CA for the certs.

[1] https://kubernetes.io/docs/reference/access-authn-authz/extensible-admission-controllers/#authenticate-apiservers